core: Resolve compatibility issues
Right of the bat we faced a few compatibility issues...

After some decision making we decided to back paddle a bit.
From now on, the core project targets the following frameworks:

.NET Standard 2.0, .NET 5.0, and .NET Framework 4.7.2.

Unfortunately, (to my understanding) we couldn't use the latest EF Core
because it has dependency on newer APIs, which means, that we would have to drop .NET Framework WPF project.

To prevent that happening, we downgraded back to EF Core 2.0.

So far so good...