Migrate to .NET 6
We are finally up to date with .NET 6!

A few breaking changes but it was worth the effort.
There is no more need to worry about compatibility issues.
I've also decided to normalize SQL identifiers to snake_case.
Why? Because it is easier to type and the final query looks cleaner
without all the noisy quotations.