(api): Secure API endpoints
When it comes to authorization, security is very important.
We had to rethink the design of the requests we are sending.
As a byproduct of refactoring the controllers are now more RESTful.
We also had to improve the error handling. It is unfortunate, that
Moodle API only returns successful, non-localized status code... not anymore!

It is still not perfect. But we are getting closer.