wpf: Realtime points calculation
It is embarrassing to admit how much I've straggled with this feature :)

Nonetheless, we can now calculate points that student receives based on activity score.
There is still plenty to do. We have to add support for state persistence when navigating between pages.

Other than that, a few quality of life improvements have been made, such as:

- Event (un)subscription has been moved from constructors to the appropriate life cycle handlers.
- Formatting fixes
- Removed unused / dead code