api: Retrieve Moodle courses
We are beginning our journey of removing hard-coded values.

First step is to acquire the actual courses alongside its assignments.
Unfortunately, I couldn't find a "convenient" way of asking for a course.
Unless a student remembers or knows the identifier of a course there is now way to get the assignments.
Of course, it's a bad user experience. It makes way more sense to get a course by a name, or its short-code.

Other than that, a few quality of life improvements have been made:

- When authorization fails => return 401 (Unauthorized) instead of 400 (Bad request)
- Default course id has been set to 5 which corresponds to our own playground.