wpf: Condense HTTP calls
Up until now the way we were calling protected HTTP endpoints was by using HTTP extensions.
Which was very awkward and unintuitive. Not anymore!

By merging HTTP extensions into BaseRequest the API now looks consistent.

Also, a minor issue I overlooked at first was the declaration of HttpClient.
We used to allocate a brand new HttpClient for every request which is an obvious performance penalty.

By marking it as static we are reusing the same HTTP instance across all requests.