Upgrade to .NET 8
Happy New Year! Due to .NET 7 being out of support it is time to upgrade to .NET 8.
Unfortunately, the transition was not smooth. There were a few breaking changes, specifically:

* EFCore.NamingConventions
* MaterialDesignThemes
* QuestPDF

could not be fully upgraded at the moment. Despite many warnings I managed to build and run the project.
We also embraced nullability. It is a way to go. "Fighting" analyzers is a death by thousand cuts.

The folder structure has been flattened to reflect the architecture. Web API is not really a server concern.

Currently, a few integration tests are failing due to ASP.NET changes. To fix them, we need to reconsider
the way we are communicating with Moodle. For now, it is a work in progress.

More to come. Better late than never.